pyproject.toml
[tool.poetry]
name = "qis"
version = "2.0.38"
description = "Implementation of visualisation and reporting analytics for Quantitative Investment Strategies"
license = "LICENSE.txt"
authors = ["Artur Sepp <artursepp@gmail.com>"]
maintainers = ["Artur Sepp <artursepp@gmail.com>"]
readme = "README.md"
repository = "https://github.com/ArturSepp/QuantInvestStrats"
documentation = "https://github.com/ArturSepp/QuantInvestStrats"
keywords= ["quantitative", "investing", "portfolio optimization", "systematic strategies", "volatility"]
classifiers=[
        "Development Status :: 4 - Beta",
        "Environment :: Console",
        "Intended Audience :: Financial and Insurance Industry",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: MIT License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3 :: Only",
        "Topic :: Office/Business :: Financial :: Investment",
    ]
packages = [ {include = "qis"} ]

[tool.poetry.urls]
"Issues" = "https://github.com/ArturSepp/QuantInvestStrats/issues"
"Personal website" = "https://artursepp.com"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numba = ">=0.56.4"
numpy = ">=1.22.4"
scipy = ">=1.10"
statsmodels = ">=0.13.5"
pandas = ">=1.5.2"
matplotlib = ">=3.2.2"
seaborn = ">=0.12.2"
openpyxl = ">=3.0.10"
tabulate = ">=0.9.0"
PyYAML = ">=6.0"
easydev = ">=0.12.0"
psycopg2 = ">=2.9.5"
SQLAlchemy = ">=1.4.46"
pyarrow = ">=10.0.1"
fsspec = ">=2022.11.0"
yfinance = "> 0.1.38"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
